test: move instanceprofile tests to its own package
