+++ pkg/handlers/aws/mutation/iaminstanceprofile/inject_suite_test.go
@@ -1,0 +1,13 @@
+package iaminstanceprofile
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestIAMInstnaceProfilePatch(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "IAMInstanceProfile patches for ControlPlane and Workers suite")
+}
