+++ pkg/handlers/aws/mutation/iaminstanceprofile/inject_worker_test.go
@@ -1,74 +1,39 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package tests
+package iaminstanceprofile
 
 import (
-	"testing"
-
+	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 
+	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/api/v1alpha1"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
+	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/workerconfig"
 )
 
-func TestControlPlaneGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
+var _ = Describe("Generate IAMInstanceProfile patches for Worker", func() {
+	patchGenerator := func() mutation.GeneratePatches {
+		return mutation.NewMetaGeneratePatchesHandler("", NewWorkerPatch()).(mutation.GeneratePatches)
+	}
 
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
+	testDefs := []capitest.PatchTestDef{
+		{
 			Name: "unset variable",
 		},
-		capitest.PatchTestDef{
-			Name: "iamInstanceProfile set",
+		{
+			Name: "iamInstanceProfile for worker set",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
-					variableName,
-					"control-plane.cluster-api-provider-aws.sigs.k8s.io",
-					variablePath...,
-				),
-			},
-			RequestItem: request.NewCPAWSMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/iamInstanceProfile",
-				ValueMatcher: gomega.Equal("control-plane.cluster-api-provider-aws.sigs.k8s.io"),
-			}},
-		},
-	)
-}
-
-func TestWorkerGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
-
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
-			Name: "unset variable",
-		},
-		capitest.PatchTestDef{
-			Name: "iamInstanceProfile set",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
+					workerconfig.MetaVariableName,
 					"nodes.cluster-api-provider-aws.sigs.k8s.io",
-					variablePath...,
+					v1alpha1.AWSVariableName,
+					VariableName,
 				),
 				capitest.VariableWithValue(
 					"builtin",
@@ -84,5 +49,17 @@ func TestWorkerGeneratePatches(
 				ValueMatcher: gomega.Equal("nodes.cluster-api-provider-aws.sigs.k8s.io"),
 			}},
 		},
-	)
-}
+	}
+
+	// create test node for each case
+	for testIdx := range testDefs {
+		tt := testDefs[testIdx]
+		It(tt.Name, func() {
+			capitest.AssertGeneratePatches(
+				GinkgoT(),
+				patchGenerator,
+				&tt,
+			)
+		})
+	}
+})
